+++ .idea/deploymentTargetSelector.xml
@@ -4,10 +4,10 @@
     <selectionStates>
       <SelectionState runConfigName="app">
         <option name="selectionMode" value="DROPDOWN" />
-        <DropdownSelection timestamp="2025-01-28T19:42:03.015609Z">
+        <DropdownSelection timestamp="2025-03-04T07:54:15.347744Z">
           <Target type="DEFAULT_BOOT">
             <handle>
-              <DeviceId pluginId="LocalEmulator" identifier="path=/Users/oskargosk/.android/avd/Pixel_8a_API_34.avd" />
+              <DeviceId pluginId="PhysicalDevice" identifier="serial=2AS0221C16001148" />
             </handle>
           </Target>
         </DropdownSelection>
